+++ inc/pinboard-bookmarks-functions-init.php
@@ -31,17 +31,21 @@ if ( ! defined( 'WPINC' ) ) {
  *    integer $truncate         Truncate the description of the bookmark at this words number.
  *    boolean $display_date     If the date of the bookmarks should be displayed.
  *    boolean $display_time     If the time of the bookmarks should be displayed.
- *    string  $date_text        The text to be prepended before the date/time.
+ *    string  $date_text        The text to be displayed before the date/time.
  *    boolean $display_tags     If the tags of the bookmarks should be displayed.
- *    boolean $tags_text        The text to be prepended before the tags.
+ *    boolean $tags_text        The text to be displayed before the tags.
  *    boolean $display_hashtag  If the hashtag `#` should be displayed.
  *    boolean $use_comma        If a comma should be displayed between tags.
  *    boolean $display_source   If the source of the bookmark should be displayed.
  *    boolean $display_arrow    If an HTML arrow should be displayed after the bookmark.
  *    integer $time             Minimum time in seconds between two requests to Pinboard server.
+ *    boolean $display_site_url If the original site base URL should be displayed.
+ *    boolean $leave_domain     If the base URL of the original should be diaplayed only.
+ *    string  $site_url_text    The text to be displayed before the original site URL.
  *    boolean $display_archive  If the link to the archive on Pinboard should be displayed.
- *    boolean $archive_text     The text to be prepended before the archive link.
- *    boolean $list_type        The HTML list type. Default 'bullet' (ul). Accepts 'numbered' (ol).
+ *    boolean $archive_text     The text to be used for the archive on Pinboard.
+ *    boolean $list_type        The HTML list type.
+ *                              Default 'bullet' (ul). Accepts 'numbered' (ol).
  *    boolean $display_arch_arr If an HTML arrow should be displayed after the archive link.
  *    boolean $new_tab          If links should be opened ina new browser tab.
  *    boolean $nofollow         If a 'nofollow' attribute should be added in links.
@@ -76,13 +80,16 @@ function pinboard_bookmarks_get_defaults() {
 		'display_source'   => false,
 		'display_arrow'    => false,
 		'time'             => 1800,
+		'display_site_url' => false,
+		'leave_domain'     => false,
+		'site_url_text'    => esc_html__( 'From:', 'pinboard-bookmarks' ),
 		'display_archive'  => true,
 		'archive_text'     => esc_html__( 'See the bookmarks on Pinboard', 'pinboard-bookmarks' ),
 		'list_type'        => 'bullet',
 		'display_arch_arr' => true,
 		'new_tab'          => false,
 		'nofollow'         => true,
-		'items_order'      => 'title description date tags',
+		'items_order'      => 'title site description date tags',
 		'admin_only'       => true,
 		'debug_options'    => false,
 		'debug_urls'       => false,